Refactor out coarsen tests (#5474)
